+++ ScriptRemote.Wpf/MainWindow.xaml
@@ -14,6 +14,7 @@
             <StackPanel Orientation="Horizontal">
                 <StackPanel.ContextMenu>
                     <ContextMenu>
+                        <MenuItem Header="Edit" Click="settingsListItem_Edit" Tag="{Binding }"></MenuItem>
                         <MenuItem Header="Delete" Click="settingsListItem_Delete" Tag="{Binding }"></MenuItem>
                     </ContextMenu>
                 </StackPanel.ContextMenu>
@@ -26,9 +27,8 @@
               Margin="2 0 0 0" DockPanel.Dock="Top" SnapsToDevicePixels="True">
             <MenuItem Header="_File">
                 <MenuItem Header="_New Session..." Click="MenuItem_Click_New" />
-                <MenuItem Header="_Reopen Session" />
                 <Separator/>
-                <MenuItem Header="E_xit" Click="MenuItem_Click_Exit" />
+                <MenuItem Header="_Exit" Click="MenuItem_Click_Exit" />
             </MenuItem>
             <MenuItem Header="_Tools">
                 <MenuItem Header="_Options" />
